Fix warnings

+++ Calibration/RCCalibration1.m
@@ -70,7 +70,7 @@
 {
     if ([status.error isKindOfClass:[RCLicenseError class]])
     {
-        NSString * message = [NSString stringWithFormat:@"There was a problem validating your license. The license error code is: %i.", status.error.code];
+        NSString * message = [NSString stringWithFormat:@"There was a problem validating your license. The license error code is: %li.", (long)status.error.code];
         UIAlertView *alert = [[UIAlertView alloc] initWithTitle:@"3DK License Error"
                                                         message:message
                                                        delegate:self
